src/requirements.txt
lxml
minio
python-dotenv
requests